dochadzka_backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
future==1.0.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
serializers==0.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
